Crate zerofrom_derive

Custom derives for ZeroFrom from the zerofrom crate.

Modules

visitor 🔒

Visitor for determining whether a type has type and non-static lifetime parameters (duplicated in yoke/derive/src/visitor.rs)

Functions

custom_lt 🔒

get_may_borrow_attr 🔒

has_attr 🔒

replace_lifetime 🔒

Replace all lifetimes in a type with a specified one

replace_lifetime_and_type 🔒

Replace all lifetimes in a type with a specified one, AND replace all types that have a corresponding C type with the C type

zf_derive_impl 🔒

Derive Macros

ZeroFrom

Custom derive for zerofrom::ZeroFrom,