Derive Macro zerofrom_derive::ZeroFrom

#[derive(ZeroFrom)]
{
    // Attributes available to this derive:
    #[zerofrom]
}

Custom derive for zerofrom::ZeroFrom,

This implements ZeroFrom<Ty> for Ty for types without a lifetime parameter, and ZeroFrom<Ty<'data>> for Ty<'static> for types with a lifetime parameter.

Apply the #[zerofrom(clone)] attribute to a field if it doesn’t implement ZeroFrom; this data will be cloned when the struct is zero_from’ed.