zerofrom_derive

Module visitor

Visitor for determining whether a type has type and non-static lifetime parameters (duplicated in yoke/derive/src/visitor.rs)

Structs

TypeVisitor 🔒

Functions

check_type_for_parameters

Checks if a type has type or lifetime parameters, given the local context of named type parameters. Returns (has_type_params, has_lifetime_params)